cloud_function/requirements.txt
blinker==1.9.0
click==8.2.1
cloudevents==1.11.1
deprecation==2.1.0
Flask==3.1.1
functions-framework==3.8.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
watchdog==6.0.0
Werkzeug==3.1.3
